tests a list of source packages

tests a list of source packages

Usage

pkgs_test(
  pkgids = names(filter_srcpkgs(src_pkgs, filter)),
  src_pkgs = get_srcpkgs(),
  filter = NULL,
  quiet = TRUE,
  ...
)

Arguments

pkgids

a list of package ids (names, paths or object), or a srcpkgs object. Also accept a singleton package object

src_pkgs

a collection of source packages as a srckgs object.

filter

filter out the packages to test using this pattern

quiet

whether to be quiet/silent

...

passed to pkg_test

Value

the results as a pkgs_test object